requirements.txt
aiobotocore==0.9.2
aiohttp==3.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
botocore==1.10.38
cachetools==2.1.0
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
cryptography==2.1.4
docutils==0.14
idna==2.7
idna-ssl==1.0.1
jmespath==0.9.3
more-itertools==4.2.0
multidict==4.3.1
pluggy==0.6.0
py==1.5.3
pycparser==2.18
PyJWT==1.5.3
pytest==3.6.1
pytest-cov==2.5.1
python-dateutil==2.7.3
PyYAML==3.12
six==1.11.0
uritemplate==3.0.0
uvloop==0.10.1
wrapt==1.10.11
yarl==1.2.5
